cine_prediction/requirements.txt
# django==4.1
# django-dotenv
scrapy==2.9.0
scikit-learn==1.2.2
Django==2.1
pyodbc==4.0.39
django-pyodbc-azure==2.1.0.*
python-dotenv==1.0.0
pandas==2.0.3
joblib==1.3.0
numpy==1.23.5
catboost==1.2
fastapi==0.95.0
uvicorn==0.23.2
gunicorn==20.1.0